pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sonolbot"
version = "0.1.0"
description = "Sonolbot Telegram daemon + CLI control layer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.1.1",
  "requests>=2.32.0",
  "click>=8.1.7",
]

[project.scripts]
sonolbot = "sonolbot.cli:main"

[project.optional-dependencies]
skills-core = [
  "defusedxml>=0.7.1",
  "lxml>=5.0.0",
]
skills-office = [
  "openpyxl>=3.1.0",
  "Pillow>=10.0.0",
]
skills-pdf = [
  "pypdf>=5.0.0",
  "pdf2image>=1.16.3",
]
skills-web = [
  "playwright>=1.44.0",
]

[project.urls]
Homepage = "https://github.com/"
